+++ native/CHANGELOG.md
@@ -6,7 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## Recent Changes
 
-- `c`: Added support to return information for current linklist.
+- `c`: Added support to return information for current linklist. [#1061](https://github.com/zowe/zowex/pull/1061)
 
 ## `0.6.1`